+++ projects/ngx-darkbox-gallery/src/lib/model/image.ts
@@ -1,22 +1,22 @@
-export class Image {
+export interface Image {
   /**
    * An id to identify the image in events. This is not used by the library itself
    */
-  public identifier?: string;
+  identifier?: string;
   /**
    * URL of the full sized image
    */
-  public url: string;
+  url: string;
   /**
    * URL of the thumbnail image
    */
-  public thumbnailUrl: string;
+  thumbnailUrl: string;
   /**
    * A short caption that can be shown under the  image in darkbox
    */
-  public caption?: string;
+  caption?: string;
   /**
    * The alternative text shown when the image can not be loaded
    */
-  public altText?: string;
+  altText?: string;
 }