+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 * Remove deprecated keyboard navigation binding
+* Convert image class to aninterface
 
 ## [5.0.1] - 2026-05-13